@@ -16,8 +16,10 @@
 
 import 'package:vector_math/vector_math_64.dart';
 
-import '../data_types.dart';
 import '../document.dart';
+import '../format/array.dart';
+import '../format/dict.dart';
+import '../format/num.dart';
 import 'font.dart';
 import 'xobject.dart';
 
