@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-import '../data_types.dart';
 import '../document.dart';
-import '../stream.dart';
+import '../format/dict_stream.dart';
+import '../format/stream.dart';
 import 'object_dict.dart';
 
 /// Stream Object
@@ -35,12 +35,12 @@ class PdfObjectStream extends PdfObjectDict {
   final bool isBinary;
 
   @override
-  void writeContent(PdfStream os) {
+  void writeContent(PdfStream s) {
     PdfDictStream.values(
-      object: this,
       isBinary: isBinary,
       values: params.values,
       data: buf.output(),
-    ).output(os, pdfDocument.verbose ? 0 : null);
+    ).output(this, s, pdfDocument.settings.verbose ? 0 : null);
+    s.putByte(0x0a);
   }
 }
