@@ -24,7 +24,7 @@ import 'package:test/test.dart';
 
 void main() {
   late Document pdf;
-  TextStyle? symbol;
+
   late ImageProvider im;
 
   setUpAll(() {
@@ -40,8 +40,6 @@ void main() {
           base: Font.ttf(defaultFont.buffer.asByteData()),
           bold: Font.ttf(defaultFontBold.buffer.asByteData()),
         ));
-
-    symbol = TextStyle(font: Font.zapfDingbats());
 
     final imData = zlib.decode(base64.decode(
         'eJz7//8/w388uOTCT6a4Ez96Q47++I+OI479mEVALyNU7z9seuNP/mAm196Ekz8YR+0dWHtBmJC9S+7/Zog89iMIKLYaHQPVJGLTD7MXpDfq+I9goNhPdPPDjv3YlnH6Jye6+2H21l/6yeB/4HsSDr1bQXrRwq8HqHcGyF6QXp9933N0tn/7Y7vn+/9gLPaih0PDlV9MIAzVm6ez7dsfzW3f/oMwzAx0e7FhoJutdbcj9MKw9frnL2J2POfBpxeEg478YLba/X0Wsl6lBXf+s0bP/s8ePXeWePJCvPEJNYMRZIYWSO/cq/9Z/Nv+M4bO+M8YDjFDJGkhzvSE7A6jRTdnsQR2wfXCMLHuMC5byyidvGgWE5JeZDOIcYdR+TpmkBno+mFmAAC+DGhl'));
@@ -85,12 +83,7 @@ void main() {
                 children: <Widget>[
                   Image(im),
                   PdfLogo(),
-                  Column(
-                    children: <Widget>[
-                      Text('(', style: symbol),
-                      Text('4', style: symbol),
-                    ],
-                  ),
+                  SizedBox.square(dimension: 20, child: FlutterLogo()),
                 ],
               ),
               Padding(
