@@ -67,8 +67,7 @@ class PdfPreviewData extends ChangeNotifier {
   PdfPageFormat get actualPageFormat => _onComputeActualPageFormat();
 
   String get localPageFormat {
-    final wbi = WidgetsBinding.instance;
-    final locale = wbi.window.locale;
+    final locale = WidgetsBinding.instance.platformDispatcher.locale;
     final cc = locale.countryCode ?? 'US';
 
     if (cc == 'US' || cc == 'CA' || cc == 'MX') {
